afsutil: avoid git clean of the wrong directory
Add some santity checks to the afsutil build command to avoid running in
a directory that does not seem to be the OpenAFS top level directory.

Add a check for a new git config option to enable the git clean before
building.  To enable the git clean, run this once in your OpenAFS git
repo;

    git config --local afsutil.clean true

Sorry for nuking your config file Joe.

+++ libraries/afsutil/afsutil/build.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2014-2015 Sine Nomine Associates
+# Copyright (c) 2014-2016 Sine Nomine Associates
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions are met:
@@ -36,9 +36,30 @@ DEFAULT_CF = [
     '--without-dot',
 ]
 
+def _sanity_check_dir():
+    msg = "Missing '%s'; are you in the OpenAFS source top-level directory?"
+    for d in ('src', 'src/afs', 'src/viced'):
+        if not os.path.isdir(d):
+            raise AssertionError(msg % (d))
+
+def _allow_git_clean():
+    clean = False
+    try:
+        output = sh('git', 'config', '--bool', '--get', 'afsutil.clean', output=True)
+        if output[0] == 'true':
+            clean = True
+    except CommandFailed as e:
+        if e.code == 1:
+            logger.info("To enable git clean before builds:")
+            logger.info("    git config --local afsutil.clean true");
+        else:
+            raise e
+    return clean
+
 def _clean():
     if os.path.isdir('.git'):
-        sh('git', 'clean', '-f', '-d', '-x', '-q')
+        if _allow_git_clean():
+            sh('git', 'clean', '-f', '-d', '-x', '-q')
     else:
         if os.path.isfile('./Makefile'):
             sh('make', 'clean')
@@ -103,6 +124,7 @@ def build(cf=None, target='all', clean=True, transarc=True, **kwargs):
     if target == 'all' and '--enable-transarc-paths' in cf:
         target = 'dest'
 
+    _sanity_check_dir()
     if clean:
         _clean()
     sh('./regen.sh')
